Small refactoring to avoid cluttering cli.py even more
