+++ senza/subcommands/root.py
@@ -1,0 +1,22 @@
+import click
+import senza
+from clickclick import AliasedGroup
+
+from ..arguments import GLOBAL_OPTIONS, region_option
+
+CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
+
+
+def print_version(ctx, param, value):
+    if not value or ctx.resilient_parsing:
+        return
+    click.echo('Senza {}'.format(senza.__version__))
+    ctx.exit()
+
+
+@click.group(cls=AliasedGroup, context_settings=CONTEXT_SETTINGS)
+@click.option('-V', '--version', is_flag=True, callback=print_version, expose_value=False, is_eager=True,
+              help='Print the current version number and exit.')
+@region_option
+def cli(region):
+    GLOBAL_OPTIONS['region'] = region
